# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Network CRM API
  description: |
    AI-powered SaaS platform API for managing personal business networks with intelligent
    relationship insights and proactive networking recommendations.

    # Authentication
    All API endpoints require authentication via Bearer token in the Authorization header.

    # Rate Limiting
    - Authenticated users: 100 requests/minute
    - AI endpoints: 10 requests/minute

    # Base URLs
    - Development: http://localhost:3000/api/v1
    - Production: https://api.pmcrm.io/v1
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@pmcrm.io
    url: https://docs.pmcrm.io
  license:
    name: Proprietary
    url: https://pmcrm.io/terms
servers:
  - url: https://api.pmcrm.io/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /contacts:
    get:
      tags:
        - Contacts
      summary: List contacts
      description: Retrieve a paginated list of contacts with optional filtering
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
        - name: relationshipStrength[gte]
          in: query
          description: Minimum relationship strength
          schema:
            type: integer
            minimum: 1
            maximum: 10
        - name: sort
          in: query
          description: Sort order (prefix with - for descending)
          schema:
            type: string
            default: -lastContactDate
            example: -lastContactDate,firstName
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Contacts
      summary: Create contact
      description: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactInput'
      responses:
        '201':
          description: Contact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /contacts/{id}:
    get:
      tags:
        - Contacts
      summary: Get contact
      description: Retrieve contact details by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Contact details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Contacts
      summary: Update contact
      description: Update contact information
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactInput'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Contacts
      summary: Delete contact
      description: Soft delete a contact
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Contact deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    LimitParam:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        default: 50
        maximum: 100
    CursorParam:
      name: cursor
      in: query
      description: Pagination cursor
      schema:
        type: string
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        linkedinUrl:
          type: string
          format: uri
        organizationId:
          type: string
          format: uuid
        relationshipStrength:
          type: integer
          minimum: 1
          maximum: 10
        lastContactDate:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContactInput:
      type: object
      required:
        - firstName
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 255
        lastName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
        linkedinUrl:
          type: string
          format: uri
        organizationId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        customFields:
          type: object
          additionalProperties: true
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: string
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            statusCode:
              type: integer
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            details:
              type: object
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...